+++ frontend2/src/index.tsx
@@ -1,22 +1,25 @@
 /* @refresh reload */
 import "virtual:uno.css"
 import { render } from 'solid-js/web'
-import { Router, Route } from "@solidjs/router";
-// import '~/index.css'
-// import App from './App.tsx'
+import { Router, Route } from "@solidjs/router"; // Import A for navigation links
 
+// Import your routes
 import Home from "~/routes/home.tsx";
 import Dashboard from "~/routes/dashboard.tsx";
+
+// Import the logo asset
+// import solidLogo from '~/assets/solid.svg';
+import { Layout } from '~/lib/components/layout';
 
 const root = document.getElementById('root')
 
 if (!root) {
   throw new Error("Wrapper div not found");
 }
-// render(() => <App />, root!)
+// Render the application
 render(
   () => (
-    <Router>
+    <Router root={Layout}>
       <Route path="/" component={Home} />
       <Route path="/dashboard" component={Dashboard} />
     </Router>